chore(api): update answer script upload functionality
