+++ services/api/.air.toml
@@ -10,7 +10,7 @@ exclude_unchanged = true
 post_cmd = ["docker stop postgres", "docker stop minio"]
 pre_cmd = [
   "docker run --rm -d --name postgres -e POSTGRES_USER=root -e POSTGRES_PASSWORD=123456 -p 5432:5432 postgres:14-alpine",
-  "docker run --rm -d --name minio -p 9000:9000 quay.io/minio/minio server /data --console-address :9001",
+  "docker run --rm -d --name minio -p 9000:9000 -p 9001:9001 quay.io/minio/minio server /data --console-address :9001",
   "sleep 3",
 ]
